C0502240900
- completed coding of exercise 140
